Separate tests into their own classes and files
